requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
alembic==1.13.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
requests==2.31.0
faker==22.5.1
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0
black==24.1.1
ruff==0.1.15
aiosqlite==0.19.0

